% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikiTools.R
\name{cc}
\alias{cc}
\title{Converts a text separated by commas into a character vector.}
\usage{
cc(text, sep = ",")
}
\arguments{
\item{text}{Text to be separated.}

\item{sep}{A character of separation. It must be a blank. If it is another character, trailing blanks are suppressed.}
}
\value{
A vector of the split segments of the text.
}
\description{
Converts a text separated by commas into a character vector.
}
\details{
Returns inside the text are omitted.
}
\examples{
## A text with three names separated with commas is converted into a vector of length 3.
cc("Pablo Picasso, Diego Velazquez, Salvador Dali")
}
\author{
Modesto Escobar, Department of Sociology and Communication, University of Salamanca. See \url{https://sociocav.usal.es/blog/modesto-escobar/}
}
