% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{checkEntities}
\alias{checkEntities}
\title{Check if all Wikidata entities in entity_list have valid values}
\usage{
checkEntities(entity_list)
}
\arguments{
\item{entity_list}{A vector with the Wikidata entities.}
}
\value{
The list of entities or raise an error.
}
\description{
Return a vector of entities with duplicates or void entities removed. A valid
entity is a wikibase item (Qxxx, x is a digit) or a wikibase property (Pxxx).
}
\author{
Angel Zazo, Department of Computer Science and Automatics, University of Salamanca
}
