% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{CheckTitles}
\alias{CheckTitles}
\title{CheckTitles(titles)
Check if titles are valid. Return TRUE is all titles are valid, else FALSE.
See:See https://en.wikipedia.org/wiki/Wikipedia:Page_name#Technical_restrictions_and_limitations}
\usage{
CheckTitles(titles)
}
\arguments{
\item{titles}{A vector of titles to check.}
}
\description{
CheckTitles(titles)
Check if titles are valid. Return TRUE is all titles are valid, else FALSE.
See:See https://en.wikipedia.org/wiki/Wikipedia:Page_name#Technical_restrictions_and_limitations
}
