% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipedia-tools.R
\name{count_refs}
\alias{count_refs}
\title{Count total <ref> tags in wikitext (includes non-templated refs)}
\usage{
count_refs(wikitext)
}
\description{
Count total <ref> tags in wikitext (includes non-templated refs)
}
