% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wikidata_instances.R
\name{.extract_instance_or_subclass}
\alias{.extract_instance_or_subclass}
\title{Extract Instance/Subclass QIDs}
\usage{
.extract_instance_or_subclass(entity, property_id = "P31")
}
\arguments{
\item{entity}{A single Wikidata entity object.}

\item{property_id}{Character. Either `"P31"` or `"P279"`.}
}
\value{
A character vector of QIDs.
}
\description{
Internal helper that extracts all P31 (instance of) or P279
  (subclass of) target QIDs from one entity.
}
\details{
Extract Instance/Subclass QIDs
}
\keyword{internal}
