% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_equivalent.R
\name{equivalent_which}
\alias{equivalent_which}
\title{Find Positions of Equivalent Strings in a Vector}
\usage{
equivalent_which(string, string_list)
}
\arguments{
\item{string}{A character string to search for.}

\item{string_list}{A character vector to search in.}
}
\value{
An integer vector of matching positions, or `integer(0)` if none
  match.
}
\description{
Returns the indices of elements in `string_list` that are equivalent to
`string` under [str_equivalent()] normalization (case, accents, whitespace,
non-breaking spaces, double quotes).
}
\examples{
equivalent_which("café", c("cafe", "tea", "coffee"))  # 1
equivalent_which("hello", c("world", "Hello"))        # 2
equivalent_which("nope", c("a", "b", "c"))            # integer(0)

}
\seealso{
[str_equivalent()], [equivalent_match()], [str_equivalent_list()]
}
