% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikiTools.R
\name{filext}
\alias{filext}
\title{Extract the extension of a file}
\usage{
filext(fn)
}
\arguments{
\item{fn}{Character vector with the files whose extensions are to be extracted.}
}
\value{
A character vector of extension names.
}
\description{
Extract the extension of a file
}
\details{
This function extracts the extension of a vector of file names.
}
\examples{
## For a single item:
filext("Albert Einstein.jpg")
## You can do the same for a vector:
filext(c("Hillary Duff.png", "Britney Spears.jpg", "Avril Lavigne.tiff"))
}
\author{
Modesto Escobar, Department of Sociology and Communication, University of Salamanca. See \url{https://sociocav.usal.es/blog/modesto-escobar/}
}
