% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipedia-tools.R
\name{get_category_pages}
\alias{get_category_pages}
\title{Get article pages in a category (non-recursive)}
\usage{
get_category_pages(category, lang = "en")
}
\description{
Get article pages in a category (non-recursive)
}
