% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipedia-tools.R
\name{get_page_info_batch}
\alias{get_page_info_batch}
\title{Fetch basic page info + Wikidata QID for a vector of titles}
\usage{
get_page_info_batch(titles, lang = "en")
}
\arguments{
\item{titles}{Character vector of page titles}

\item{lang}{Language code}
}
\value{
tibble with: title, pageid, length, wikidata_qid
}
\description{
Fetch basic page info + Wikidata QID for a vector of titles
}
