% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipedia-tools.R
\name{get_wikitext_from_url}
\alias{get_wikitext_from_url}
\title{Fetch wikitext from a full Wikipedia URL (standard or old-revision)}
\usage{
get_wikitext_from_url(url)
}
\description{
Fetch wikitext from a full Wikipedia URL (standard or old-revision)
}
