% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipedia-tools.R
\name{get_wp_category_pages}
\alias{get_wp_category_pages}
\title{Get Pages from a Category}
\usage{
get_wp_category_pages(category, lang = "en")
}
\arguments{
\item{category}{Character. Category name.}

\item{lang}{Character. Wikipedia language code.}
}
\value{
A tibble with page members.
}
\description{
Convenience wrapper around `get_category_members()` for pages
  only (non-recursive).
}
\details{
Get Pages from a Category
}
