% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikiTools.R
\name{getWikiData}
\alias{getWikiData}
\title{Create a data.frame with Wikidata of a vector of names.}
\usage{
getWikiData(names, csv = NULL)
}
\arguments{
\item{names}{A vector consisting of one or more Wikidata's entry (i.e., topic or person).}

\item{csv}{A file name to save the results, in which case the only return is a message with the name of the saved file.}
}
\value{
A data frame with personal information of the names or a csv file with the information separated by semicolons.
}
\description{
Create a data.frame with Wikidata of a vector of names.
}
\examples{
## Obtaining information in English Wikidata
names <- c("Douglas Adams", "Picasso")
information <- getWikiData(names)

## Obtaining information in Spanish Wikidata
informacion <- getWikiData(names)
}
\author{
Modesto Escobar, Department of Sociology and Communication, University of Salamanca. See \url{https://sociocav.usal.es/blog/modesto-escobar/}
}
