% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikiTools.R
\name{getWikiQ}
\alias{getWikiQ}
\title{Create a data.frame with Q's and descriptions of a vector of names.}
\usage{
getWikiQ(names, number = 1, language = "en")
}
\arguments{
\item{names}{A vector consisting of one or more Wikidata's entry (i.e., topic or person).}

\item{number}{Take the number occurrence in case there are several equal names in Wikidata.}

\item{language}{The language of the Wikipedia page version. This should consist of an ISO language code (default = "en").}
}
\value{
A data frame with name, Q, label and description of the names.
}
\description{
Create a data.frame with Q's and descriptions of a vector of names.
}
\examples{
## Obtaining information in English Wikidata
names <- c("Douglas Adams", "Picasso")
information <- getWikiQ(names)

## Obtaining information in Spanish Wikidata
informacion <- getWikiQ(names)
}
