% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{httrGetJSON}
\alias{httrGetJSON}
\title{httrGetJSON
Retrieve responses in JSON format using httr::GET. It is a generic function
to use for request these Wikimedia metrics API:
https://wikimedia.org/api/rest_v1/
https://www.mediawiki.org/wiki/XTools/API/Page (xtools.wmflabs.org)}
\usage{
httrGetJSON(url)
}
\arguments{
\item{url}{The URL with the query to the API.}
}
\value{
A JSON response. Please check httr::stop_for_status(response)
}
\description{
httrGetJSON
Retrieve responses in JSON format using httr::GET. It is a generic function
to use for request these Wikimedia metrics API:
https://wikimedia.org/api/rest_v1/
https://www.mediawiki.org/wiki/XTools/API/Page (xtools.wmflabs.org)
}
\note{
Used in m_Pageviews
}
\author{
Angel Zazo, Department of Computer Science and Automatics, University of Salamanca
}
