% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{limit_requester}
\alias{limit_requester}
\title{limit_requester
Limit the rate at which a function will execute}
\usage{
limit_requester(f, n, period)
}
\arguments{
\item{f}{The original function}

\item{n}{Number of allowed events within a period}

\item{period}{Length (in seconds) of measurement period}
}
\value{
If 'f' is a single function, then a new function with the same signature and (eventual) behavior as the original function, but rate limited. If 'f' is a named list of functions, then a new list of functions with the same names and signatures, but collectively bound by a shared rate limit.
}
\description{
limit_requester
Limit the rate at which a function will execute
}
\seealso{
ratelimitr
}
\author{
Angel F. Zazo, Departament of Computer Science and Automatics, University of Salamanca
}
