% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wikitable.R
\name{link_to_article_name}
\alias{link_to_article_name}
\title{Extract a Wikipedia Article Name from a Link}
\usage{
link_to_article_name(link)
}
\arguments{
\item{link}{Character vector of URLs/hrefs, such as the `_link` columns
added by [get_wikitable()].}
}
\value{
Character vector of article titles, same length as `link`.
}
\description{
Parses a Wikipedia article URL (or `href`, absolute or relative) and
returns the plain article title, with underscores replaced by spaces.
Values that don't contain `"/wiki/"` (including `NA`) return `NA`.
}
\examples{
link_to_article_name("https://en.wikipedia.org/wiki/Please_Please_Me")
link_to_article_name("/wiki/A_Hard_Day\%27s_Night")
link_to_article_name(c("https://en.wikipedia.org/wiki/Help!", NA, "not a link"))

}
\seealso{
[get_wikitable()], [clean_column_headers()]
}
