% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{m_WikidataEntity}
\alias{m_WikidataEntity}
\title{Check Wikipedia pages titles}
\usage{
m_WikidataEntity(titles, project = "en.wikipedia.org", redirects = TRUE)
}
\arguments{
\item{titles}{A vector of page titles to search for.}

\item{project}{Wikimedia project, defaults "en.wikipedia.org"}

\item{redirects}{If page redirects must be resolved. If redirects=TRUE
(default) then the "normalized" column of the returned data-frames contains
the destiny page title of the redirection.}
}
\value{
Null if there is any error in response, else a data-frame with four
columns: first, the original page title string, second, the normalized one,
third, logical error=FALSE, if Wikidata entity exists for the page, or
error=TRUE it does not, last, the Wikidata entity itself or a clarification
of the error.
}
\description{
Use reqMediaWiki to check if page titles are in a Wikimedia project and returns
the Wikidata entity for them. Automatically resolves redirects if parameter
redirects = TRUE (default). If a page title exists in the Wikimedia project,
the status column in the returned data-frame is set to 'OK'. If a page is a
disambiguation page, that column is set to 'disambiguation', and if a title
is not in the Wikimedia project, it is set to 'missing' and no Wikidata
entity is returned.
}
\note{
It is not expected that there will be "continue" type responses,
nevertheless, their handling has been included in the code.
}
\examples{
# Note that URLdecode("a\%CC\%8C") is
# the letter "a" with the combining caron
df <- m_WikidataEntity(c('Max Planck', URLdecode("a\%CC\%8C"), 'Max', 'Cervante', 'humanist'),
                        project='en.wikipedia.org')
}
\author{
Angel Zazo, Department of Computer Science and Automatics, University of Salamanca
}
