% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{m_XtoolsInfoAll}
\alias{m_XtoolsInfoAll}
\title{m_XtoolsInfoAll(article, project="en.wikipedia.org", redirects=FALSE)
Obtains information about a page in the Wikimedia project using wmflabs
in JSON format, or NULL on error. The XTools Page API endpoints offer data
related to a single page. See https://www.mediawiki.org/wiki/XTools/API/Page
Obtains information of articleinfo, prose and links about a page in
the WikiMedia project.}
\usage{
m_XtoolsInfoAll(article, project = "en.wikipedia.org", redirects = FALSE)
}
\arguments{
\item{article}{The title of the article to search. Only one article is allowed.}

\item{project}{The Wikimedia project, defaults en.wikipedia.org}

\item{redirects}{If redirects==TRUE, then the information is obtained
of the destiny of the page. In that case, then the "original" element of
the returned list contains the original page, and the "page" element the
destiny page. Also, if infotype=='links, the sum of the in-links of all
redirections is assigned to links_in_count.}
}
\value{
The information about the page.
Note: use the m_XtoolsInfo function.
}
\description{
m_XtoolsInfoAll(article, project="en.wikipedia.org", redirects=FALSE)
Obtains information about a page in the Wikimedia project using wmflabs
in JSON format, or NULL on error. The XTools Page API endpoints offer data
related to a single page. See https://www.mediawiki.org/wiki/XTools/API/Page
Obtains information of articleinfo, prose and links about a page in
the WikiMedia project.
}
\examples{
\dontrun{
z  <- m_XtoolsInfoAll(article="Miguel de Cervantes", project="es.wikipedia.org")
zz <- m_XtoolsInfoAll(article="Cervantes", project="es.wikipedia.org",
                       redirects=TRUE)
}
}
