% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{MW_GetImages}
\alias{MW_GetImages}
\title{MW_GetImages(titles, project, exclude_ext)
Search for URL images (files) in the Wikipedia pages. Exclude extensions in
exclude_ext. See https://en.wikipedia.org/w/api.php?action=help&modules=query\%2Bimages}
\usage{
MW_GetImages(
  titles,
  project = "en.wikipedia.org",
  exclude_ext = "svg|webp|xcf"
)
}
\arguments{
\item{titles}{A vector of article titles to search for.}

\item{project}{Wikimedia project, defaults "en.wikipedio.org".}

\item{exclude_ext}{extensions excluded in results. Default 'svg|webp|xcf'}
}
\value{
A data-frame with original titles, normalized ones, logical TRUE or
FALSE and the images of wikipedia page (use "|" to separate ones) or NA if
images do not exits.
}
\description{
MW_GetImages(titles, project, exclude_ext)
Search for URL images (files) in the Wikipedia pages. Exclude extensions in
exclude_ext. See https://en.wikipedia.org/w/api.php?action=help&modules=query\%2Bimages
}
\note{
Only for namespace 0.
}
\author{
Angel Zazo, Department of Computer Science and Automatics, University of Salamanca
}
