% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{MW_GetPrimaryImage}
\alias{MW_GetPrimaryImage}
\title{MW_GetPrimaryImage(titles, project)
Return URL of image associated with the Wikipedia pages, if any.
See https://www.mediawiki.org/w/api.php?action=help&modules=query\%2Bpageimages}
\usage{
MW_GetPrimaryImage(titles, project = "en.wikipedia.org")
}
\arguments{
\item{titles}{A vector of article titles to search for.}

\item{project}{Wikimedia project, defaults "en.wikipedio.org".}
}
\value{
A data-frame with original titles, normlized ones, logical TRUE or
FALSE and the primary image of wikipedia page or NA if does not exits.
}
\description{
MW_GetPrimaryImage(titles, project)
Return URL of image associated with the Wikipedia pages, if any.
See https://www.mediawiki.org/w/api.php?action=help&modules=query\%2Bpageimages
}
\note{
Only for namespace 0.
}
\author{
Angel Zazo, Department of Computer Science and Automatics, University of Salamanca
}
