% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{MW_GetRedirects}
\alias{MW_GetRedirects}
\title{MW_GetRedirects(titles, project)
Obtains the redirection pages (from namespace 0) to the article titles in
the Wikimedia project.}
\usage{
MW_GetRedirects(titles, project = "en.wikipedia.org")
}
\arguments{
\item{titles}{A vector of article titles to search for.}

\item{project}{Wikimedia project, defaults "en.wikipedio.org"}
}
\value{
A list whose names are the titles searched and elements are all
redirections to the first element (which is the normalized title).
}
\description{
MW_GetRedirects(titles, project)
Obtains the redirection pages (from namespace 0) to the article titles in
the Wikimedia project.
}
\note{
Only for namespace 0. The function also obtains redirections for disambiguation pages.
}
\author{
Angel Zazo, Department of Computer Science and Automatics, University of Salamanca
}
