% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{MW_GetWikidataItem}
\alias{MW_GetWikidataItem}
\title{MW_GetWikidataItem(titles, project, resolv_redirects)
Use MediaWiki_query to obtain the Wikidata entity of a article title from a Wikimedia project.
Automatically resolves redirects if parameter resolv_redirects = TRUE (default).}
\usage{
MW_GetWikidataItem(
  titles,
  project = "en.wikipedia.org",
  resolv_redirects = TRUE
)
}
\arguments{
\item{titles}{A vector of article titles to search for.}

\item{project}{Wikimedia project, defaults "en.wikipedia.org"}

\item{resolv_redirects}{If article redirects must be resolved (defaults TRUE)}
}
\value{
Null if errors in response, else a data.frame with four columns:
first, the original article string, second, the normalized one, third,
logical FALSE or TRUE if the wikidata entity exists, last, the wikidata
entity itself or a clarification of the error.
}
\description{
MW_GetWikidataItem(titles, project, resolv_redirects)
Use MediaWiki_query to obtain the Wikidata entity of a article title from a Wikimedia project.
Automatically resolves redirects if parameter resolv_redirects = TRUE (default).
}
\note{
It is not expected that there will be "continue" type responses,
nevertheless, their handling has been included in the code.
}
\examples{
MW_GetWikidataItem('Max Planck', project='es.wikipedia.org')

# Note that "a\%CC\%8C" is an "a" + combining caron.
MW_GetWikidataItem(c(URLdecode("a\%CC\%8C"), 'Max', 'Cervante', 'humanist'))
}
\author{
Angel Zazo, Department of Computer Science and Automatics, University of Salamanca
}
