% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{MW_Opensearch}
\alias{MW_Opensearch}
\title{MW_Opensearch(string, project, profile, redirects)
Search y string in article title using OpenSearch. Only in namespace 0.}
\usage{
MW_Opensearch(
  string,
  project = "en.wikipedia.org",
  profile = "engine_autoselect",
  redirects = "resolve"
)
}
\arguments{
\item{string}{String to search.}

\item{project}{Wikimedia project, defaults "en.wikipedio.org".}

\item{profile}{See: https://www.mediawiki.org/wiki/API:Opensearch. This
parameter sets the search type: strict, normal, fuzzy, fast-fuzzy, classic
and engine_autoselect (default).}

\item{redirects}{See: https://www.mediawiki.org/wiki/API:Opensearch. If
redirects='return', the title is the normalized one (also the url). If
redirets='resolve", the title is the normalized and resolved redirection
is in effect (also the url). Note that in both cases the API performs a NFC
Unicode normalization on search string.}
}
\value{
A data-frame of titles and URL returned. If error, return Null.
}
\description{
MW_Opensearch(string, project, profile, redirects)
Search y string in article title using OpenSearch. Only in namespace 0.
}
\note{
Only for namespace 0. The function also obtains redirections for disambiguation pages.
}
\author{
Angel Zazo, Department of Computer Science and Automatics, University of Salamanca
}
