% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{MW_PagePrimaryImage}
\alias{MW_PagePrimaryImage}
\title{MW_PagePrimaryImage(titles, project="en.wikipedia.org")
Return the URL of the image associated with the Wikipedia pages of the titles,
if pages has one. Automatically resolves redirects, the "normalized" column
of the returned data-frames contains the destiny page of the redirection.
See https://www.mediawiki.org/w/api.php?action=help&modules=query\%2Bpageimages}
\usage{
MW_PagePrimaryImage(titles, project = "en.wikipedia.org")
}
\arguments{
\item{titles}{A vector of page titles to search for.}

\item{project}{Wikimedia project, default "en.wikipedio.org".}
}
\value{
A data-frame with original titles, normalized ones, the status of the
pages and the primary image of the page or NA if it does not exist.
}
\description{
MW_PagePrimaryImage(titles, project="en.wikipedia.org")
Return the URL of the image associated with the Wikipedia pages of the titles,
if pages has one. Automatically resolves redirects, the "normalized" column
of the returned data-frames contains the destiny page of the redirection.
See https://www.mediawiki.org/w/api.php?action=help&modules=query\%2Bpageimages
}
\examples{
i <- MW_PagePrimaryImage(c('Max Planck', URLdecode("a\%CC\%8C"), 'Max', 'Cervante', 'humanist'))
}
\author{
Angel Zazo, Department of Computer Science and Automatics, University of Salamanca
}
