% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{MW_XtoolsInfo}
\alias{MW_XtoolsInfo}
\title{MW_XtoolsInfo
Obtains information about an article in the Wikimedia project using wmflabs
in JSON format, or NULL on error. The XTools Page API endpoints offer data
related to a single page. See https://www.mediawiki.org/wiki/XTools/API/Page
[xtools.wmflabs.org]}
\usage{
MW_XtoolsInfo(
  article,
  infotype = "articleinfo",
  project = "en.wikipedia.org",
  inlinks_redirects = FALSE
)
}
\arguments{
\item{article}{The title of the article to search}

\item{infotype}{The type of information to request: articleinfo, prose, links}

\item{project}{The Wikimedia project, defaults en.wikipedia.org}

\item{inlinks_redirects}{If infotype==links, if redirects are included or not}
}
\value{
The information about the article.
}
\description{
MW_XtoolsInfo
Obtains information about an article in the Wikimedia project using wmflabs
in JSON format, or NULL on error. The XTools Page API endpoints offer data
related to a single page. See https://www.mediawiki.org/wiki/XTools/API/Page
[xtools.wmflabs.org]
}
\note{
Is important that the article don't be a redirection: if infotype=prose
the function gets information of the target article, but with infotype=articleinfo
or infotype=links the information is about the redirection.
}
\author{
Angel Zazo, Department of Computer Science and Automatics, University of Salamanca
}
