% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{normalizedTitle}
\alias{normalizedTitle}
\title{normalizedTitle(title, q)
Return de normalized or redirect title (also normalized) from the query part
of the JSON response of a MediaWiki search that uses titles.}
\usage{
normalizedTitle(title, q)
}
\arguments{
\item{title}{the title to possibly found in q.}

\item{q}{(j$query) Query part of the JSON response from a Mediawiki search.}
}
\value{
The normalized or redirect title found in q for title, else title
itself.
}
\description{
normalizedTitle(title, q)
Return de normalized or redirect title (also normalized) from the query part
of the JSON response of a MediaWiki search that uses titles.
}
