% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikiTools.R
\name{preName}
\alias{preName}
\title{Reverse the order of the first and last names of every element of a vector.}
\usage{
preName(X)
}
\arguments{
\item{X}{A vector of names with format "name, prename".}
}
\value{
Another vector with its elements changed.
}
\description{
Reverse the order of the first and last names of every element of a vector.
}
\details{
This function reverses the order of the first and last names of the items: i.e., "Weber, Max" turns into "Max Weber".
}
\examples{
## To reconvert a single name:
preName("Weber, Max")
## It is possible to work with several items, as in here:
A <- c("Weber, Max", "Descartes, Rene", "Locke, John")
preName(A)
}
\author{
Modesto Escobar, Department of Sociology and Communication, University of Salamanca. See \url{https://sociocav.usal.es/blog/modesto-escobar/}
}
