% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikiTools.R
\name{readFile}
\alias{readFile}
\title{Reads a plain text file.}
\usage{
readFile(file, encoding = "UTF-8")
}
\arguments{
\item{file}{Path and name of the file to be read.}

\item{encoding}{A character string describing the encoding format, i.e. UTF-8.}
}
\value{
Returns a character string.
}
\description{
Reads a plain text file.
}
\details{
This function allows to read a plain text file.
}
\examples{
webfile<- "https://sociocav.usal.es/me/file.txt"

file <- readFile(file = webfile)
}
\author{
Modesto Escobar, Department of Sociology and Communication, University of Salamanca. See \url{https://sociocav.usal.es/blog/modesto-escobar/}
}
