% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{req_WDQS}
\alias{req_WDQS}
\title{req_WDQS
Retrieve responses in JSON format from Wikidata Query Service (WDQS)}
\usage{
req_WDQS(sparql_query)
}
\arguments{
\item{sparql_query}{A string with the query in SPARQL language.}
}
\value{
A JSON response. Please check httr::stop_for_status(response)
}
\description{
req_WDQS
Retrieve responses in JSON format from Wikidata Query Service (WDQS)
}
\note{
For short queries GET method is better, POST for long ones. Only GET queries as cached.
}
\author{
Angel F. Zazo, Departament of Computer Science and Automatics, University of Salamanca
}
