% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{req_WDQS_rated}
\alias{req_WDQS_rated}
\title{req_WDQS_rated
The ratelimitr version of req_WDQS_rate, See
See https://www.mediawiki.org/wiki/Wikidata_Query_Service/User_Manual#SPARQL_endpoint}
\usage{
req_WDQS_rated(sparql_query)
}
\description{
req_WDQS_rated
The ratelimitr version of req_WDQS_rate, See
See https://www.mediawiki.org/wiki/Wikidata_Query_Service/User_Manual#SPARQL_endpoint
}
