% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{req_wikimedia_metrics}
\alias{req_wikimedia_metrics}
\title{req_wikimedia_metrics
Retrieve responses in JSON format from Wikimedia metrics API:
https://wikimedia.org/api/rest_v1/
https://www.mediawiki.org/wiki/XTools/API/Page [xtools.wmflabs.org]}
\usage{
req_wikimedia_metrics(url)
}
\arguments{
\item{url}{The URL with the query}
}
\value{
A JSON response. Please check httr::stop_for_status(response)
}
\description{
req_wikimedia_metrics
Retrieve responses in JSON format from Wikimedia metrics API:
https://wikimedia.org/api/rest_v1/
https://www.mediawiki.org/wiki/XTools/API/Page [xtools.wmflabs.org]
}
\note{
Used in WM_Pageviews
}
\author{
Angel Zazo, Department of Computer Science and Automatics, University of Salamanca
}
