% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{req_wikimedia_metrics_rated}
\alias{req_wikimedia_metrics_rated}
\title{req_wikimeida_metrics_rated
The limitratedr version of req_wikimedia_metrics.}
\usage{
req_wikimedia_metrics_rated(url)
}
\description{
req_wikimeida_metrics_rated
The limitratedr version of req_wikimedia_metrics.
}
