% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{v_GetRecord}
\alias{v_GetRecord}
\title{Gets record clusters using viafID or other identifier in the VIAF cluster
record.}
\usage{
v_GetRecord(sid, source = "VIAF")
}
\arguments{
\item{sid}{The identifier in the authority source.}

\item{source}{The code of the authority source in VIAF (i.e, BNF, BNE, ISNI...).}
}
\value{
The VIAF record cluster in the JSON format.
}
\description{
Obtains the record cluster in JSON format. Other formats exists, but the more
informative is 'application/json', so it is the only format used in this
function. See https://developer.api.oclc.org/viaf-api
}
\details{
If source='VIAF' (default) the function uses the "Read by ID" interface of
the API. If other authority source is used, then the function use the
localAuthorityId/authoritySourceCode interface.
See https://developer.api.oclc.org/viaf-api

Note that the returned record may be a VIAF cluster record or a
redirect/scavenged record: the function returns the record as is.
}
\examples{
\dontrun{
a <- v_GetRecord('29550309')
b <- v_GetRecord('0000000059425637', source = 'ISNI')
c <- v_GetRecord('02853882X', source = 'SUDOC')
identical(a,b)
identical(b,c)
}
}
