% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{v_Occupations}
\alias{v_Occupations}
\title{Gets occupations from a VIAF record}
\usage{
v_Occupations(viaf)
}
\arguments{
\item{viaf}{VIAF cluster record (in JSON format).}
}
\value{
A data-frame with sources and occupations from each source or NULL if
occupations do not exist in the record.
}
\description{
Return the occupations from the VIAF record. Note that the VIAF record musts
be in JSON format.
}
