% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{v_Search_allTitle}
\alias{v_Search_allTitle}
\title{v_Search_allTitle(title)
This function is a wrapper to v_Search, using this CQL_Query:
'local.title all "title"'
Search for titles.
The search is performed with the "all" operator (for all the listed terms).}
\usage{
v_Search_allTitle(name)
}
\arguments{
\item{name}{It is the title o search.}
}
\value{
A list with the records found.
}
\description{
v_Search_allTitle(title)
This function is a wrapper to v_Search, using this CQL_Query:
'local.title all "title"'
Search for titles.
The search is performed with the "all" operator (for all the listed terms).
}
