% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{v_Wikipedias}
\alias{v_Wikipedias}
\title{Gets Wikipedia pages (URL) from a VIAF record}
\usage{
v_Wikipedias(viaf)
}
\arguments{
\item{viaf}{VIAF cluster record (in JSON format).}
}
\value{
A vector with the URL of the Wikipedias.
}
\description{
Return the Wikipedia pages (URL) from the VIAF record. Note that the VIAF
record musts be in JSON format.
}
