% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikiTools.R
\name{validUrl}
\alias{validUrl}
\title{Find if an URL link is valid.}
\usage{
validUrl(url, time = 2)
}
\arguments{
\item{url}{A vector of URLs.}

\item{time}{The timeout (in seconds) to be used for each connection. Default = 2.}
}
\value{
A boolean value of TRUE or FALSE.
}
\description{
Find if an URL link is valid.
}
\details{
This function checks if a URL exists on the Internet.
}
\examples{
validUrl(url="https://es.wikipedia.org/wiki/Weber,_Max", time=2)
}
\author{
Modesto Escobar, Department of Sociology and Communication, University of Salamanca. See \url{https://sociocav.usal.es/blog/modesto-escobar/}
}
