% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{VIAF_AutoSuggest}
\alias{VIAF_AutoSuggest}
\title{VIAF_AutoSuggest(name)
Search the name string from the VIAF AutoSuggest API and returns
information in JSON format of the records found. Note that only
returns a maximum of 10 records. Note that those records are not
VIAF cluster records.}
\usage{
VIAF_AutoSuggest(name)
}
\arguments{
\item{name}{string to search}
}
\value{
A list with the records returned.
}
\description{
VIAF_AutoSuggest(name)
Search the name string from the VIAF AutoSuggest API and returns
information in JSON format of the records found. Note that only
returns a maximum of 10 records. Note that those records are not
VIAF cluster records.
}
\note{
Note the structure of the string name in the search:
auto: last name, first name[,] [([year_of_bird][-year_of_death])]
}
\seealso{
https://www.oclc.org/developer/api/oclc-apis/viaf/authority-cluster.en.html
}
