% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{VIAF_Dates}
\alias{VIAF_Dates}
\title{VIAF_Dates(viaf)
Returns bird year and death year from the VIAF cluster record with this
format byear:dyear. Note that the VIAF record musts be in JSON format.}
\usage{
VIAF_Dates(viaf)
}
\arguments{
\item{viaf}{VIAF cluster record (in JSON format).}
}
\value{
The bird year and death year in format byear:dyear.
}
\description{
VIAF_Dates(viaf)
Returns bird year and death year from the VIAF cluster record with this
format byear:dyear. Note that the VIAF record musts be in JSON format.
}
