% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{VIAF_Gender}
\alias{VIAF_Gender}
\title{VIAF_Gender(viaf)
Return the gender of the author from the VIAF record. Note that the VIAF
record musts be in JSON format.}
\usage{
VIAF_Gender(viaf)
}
\arguments{
\item{viaf}{VIAF cluster record (in JSON format).}
}
\value{
The gender of the author o NULL if not exits in the record.
}
\description{
VIAF_Gender(viaf)
Return the gender of the author from the VIAF record. Note that the VIAF
record musts be in JSON format.
}
