% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{VIAF_Occupations}
\alias{VIAF_Occupations}
\title{VIAF_Occupations(viaf)
Return the occupations from the VIAF record. Note that the VIAF record musts
be in JSON format.}
\usage{
VIAF_Occupations(viaf)
}
\arguments{
\item{viaf}{VIAF cluster record (in JSON format).}
}
\value{
A data-frame with sources and occupations from each source or NULL if
occupations do not exist in the record.
}
\description{
VIAF_Occupations(viaf)
Return the occupations from the VIAF record. Note that the VIAF record musts
be in JSON format.
}
