% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{VIAF_Search}
\alias{VIAF_Search}
\title{VIAF_Search(CQL_Query)
Run the CQL_Query using the VIAF Search API and return a list of records
found. The search string is formed using the CQL_Query syntax of the API.
Note that returned records use the "info:srw/schema/1/JSON" record schema,
i.e., are complete cluster records packed in JSON format. If the number
of records found is greater than 250 (API restrictions), successive requests
are made.}
\usage{
VIAF_Search(CQL_Query)
}
\arguments{
\item{CQL_Query}{String with the search.
See https://www.oclc.org/developer/api/oclc-apis/viaf/authority-cluster.en.html}
}
\value{
A list with the records found.
}
\description{
VIAF_Search(CQL_Query)
Run the CQL_Query using the VIAF Search API and return a list of records
found. The search string is formed using the CQL_Query syntax of the API.
Note that returned records use the "info:srw/schema/1/JSON" record schema,
i.e., are complete cluster records packed in JSON format. If the number
of records found is greater than 250 (API restrictions), successive requests
are made.
}
\examples{
\dontrun{
## Search in any field (cql.any)
# Operator is "=": so search one or more terms:
CQL_Query <- 'cql.any = "García Iranzo, Juan"'
r <- VIAF_Search(CQL_Query)
# r contains complete VIAF records (sometimes seen as a "cluster record",
# which is unified by combining records from many libraries around the world)
# Search in 1xx, 4xx, 5xx fields of MARC record (local.names)
# Operator is "all": search all terms
CQL_Query <- 'local.names all "Modesto Escobar"'
r <- VIAF_Search(CQL_Query)

# Search in 100, 400, 500 fields of MARC record (local.personalNames)
# Operator is "all": search all terms
CQL_Query <- 'local.personalNames all "Modesto Escobar"'
r <- VIAF_Search(CQL_Query)

# Search in Titles
CQL_Query <- 'local.title all "Los pronósticos electorales con encuestas"'
r <- VIAF_Search(CQL_Query)
}
}
