% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{VIAF_Search_allmainHeadingEl}
\alias{VIAF_Search_allmainHeadingEl}
\title{VIAF_Search_allmainHeadingEl(name)
This function is a wrapper to VIAF_Search, using this CQL_Query:
'local.mainHeadingEl all "name"'
Search preferred Name - names which are the preferred form in an authority
record (1xx fields of the MARC records).
1xx: Main Entry: Personal name (100), corporate name (110), meeting name (111), uniform title (130)
The search is performed with the "all" operator (all terms are searched).}
\usage{
VIAF_Search_allmainHeadingEl(name)
}
\arguments{
\item{name}{It is the name o search.}
}
\value{
A list with the records found.
}
\description{
VIAF_Search_allmainHeadingEl(name)
This function is a wrapper to VIAF_Search, using this CQL_Query:
'local.mainHeadingEl all "name"'
Search preferred Name - names which are the preferred form in an authority
record (1xx fields of the MARC records).
1xx: Main Entry: Personal name (100), corporate name (110), meeting name (111), uniform title (130)
The search is performed with the "all" operator (all terms are searched).
}
\note{
Note that the search is performed with the "all" operator, it is not
an "=" search. The "=" operator searches for all words and only those words.
With the "all" operator, more words can appear in the field.
}
