% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{VIAF_Search_allNames}
\alias{VIAF_Search_allNames}
\title{VIAF_Search_allNames(name)
This function is a wrapper to VIAF_Search, using this CQL_Query:
'local.names all "name"'
Search Names - any name preferred or alternate 1xx, 4xx, 5xx fields of the
MARC records):
1xx: Main Entry: Personal name (100), corporate name (110), meeting name (111), uniform title (130)
4xx: Series entry: Personal name (400), corporate name (410), meeting name (411), uniform title (440 - obsolete)
5xx: Note Fields
The search is performed with the "all" operator (for all the listed terms).}
\usage{
VIAF_Search_allNames(name)
}
\arguments{
\item{name}{It is the name o search.}
}
\value{
A list with the records found.
}
\description{
VIAF_Search_allNames(name)
This function is a wrapper to VIAF_Search, using this CQL_Query:
'local.names all "name"'
Search Names - any name preferred or alternate 1xx, 4xx, 5xx fields of the
MARC records):
1xx: Main Entry: Personal name (100), corporate name (110), meeting name (111), uniform title (130)
4xx: Series entry: Personal name (400), corporate name (410), meeting name (411), uniform title (440 - obsolete)
5xx: Note Fields
The search is performed with the "all" operator (for all the listed terms).
}
