% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{VIAF_Search_allPersonalNames}
\alias{VIAF_Search_allPersonalNames}
\title{VIAF_Search_allPersonalNames(name)
This function is a wrapper to VIAF_Search, using this CQL_Query:
'local.personalNames all "name"'
Search Personal Names within the authority record (100, 400, 500 fields of
MARC records):
100: Main Entry: Personal name
400: Series entry: Personal name (400)
500: Note field: General Note.
The search is performed with the "all" operator (for all the listed terms).}
\usage{
VIAF_Search_allPersonalNames(name)
}
\arguments{
\item{name}{It is the name o search.}
}
\value{
A list with the records found.
}
\description{
VIAF_Search_allPersonalNames(name)
This function is a wrapper to VIAF_Search, using this CQL_Query:
'local.personalNames all "name"'
Search Personal Names within the authority record (100, 400, 500 fields of
MARC records):
100: Main Entry: Personal name
400: Series entry: Personal name (400)
500: Note field: General Note.
The search is performed with the "all" operator (for all the listed terms).
}
