% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{VIAF_SourceId}
\alias{VIAF_SourceId}
\title{VIAF_SourceId(viaf, source)
Return the text and identifier of the source from the VIAF record.
Note that the VIAF record musts be in JSON format.}
\usage{
VIAF_SourceId(viaf, source)
}
\arguments{
\item{viaf}{VIAF cluster record (in JSON format).}

\item{source}{the identifier of the source (LC, WKP, JPG, BNE...)}
}
\value{
A data-frame with columns text and source, or NULL if the source does
no exist in the viaf record.
}
\description{
VIAF_SourceId(viaf, source)
Return the text and identifier of the source from the VIAF record.
Note that the VIAF record musts be in JSON format.
}
