% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{VIAF_Wikipedias}
\alias{VIAF_Wikipedias}
\title{VIAF_Wikipedias(viaf)
Return the Wikipedia pages (URL) from the VIAF record. Note that the VIAF
record musts be in JSON format.}
\usage{
VIAF_Wikipedias(viaf)
}
\arguments{
\item{viaf}{VIAF cluster record (in JSON format).}
}
\value{
A vector with the URL of the Wikipedias.
}
\description{
VIAF_Wikipedias(viaf)
Return the Wikipedia pages (URL) from the VIAF record. Note that the VIAF
record musts be in JSON format.
}
