% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{w_EntityInfo_tiny}
\alias{w_EntityInfo_tiny}
\title{Get a few personal properties from a Wikidata entity}
\usage{
w_EntityInfo_tiny(entity, langsorder = "en", wikilangs = "")
}
\arguments{
\item{entity}{The Wikidata entity to search for properties.}

\item{langsorder}{Order of languages in which the information will be
returned, separated with '|'. If no information is given in the first
language, next is used. This parameter is mandatory, at least one language is
required, default, 'en'.
Note: sometimes not label in any language of langsorder is assigned to an
entity, so an additional search is used to obtain almost one label for it
(?entitylab) with LIMIT 1.}

\item{wikilangs}{List of languages to limit the search of Wikipedia pages,
using "|" as separator. Wikipedias pages are returned in same order as
languages in this parameter. If wikilangs='' the function returns Wikipedia
pages in any language, not sorted.}
}
\value{
: A data-frame with the properties of entity.
}
\description{
Same as w_EntityInfo function but less properties are requested and
less checks are done.
Gets some properties of the Wikidata "entity" related to birth and death
dates, places, occupations, works, education, awards, identifier in some
libraries, Wikipedia page titles (which can be limited to the languages in
the "wikilangs" parameter), etc.
}
\examples{
df2 <- w_EntityInfo_tiny(entity='Q134644', langsorder = 'es|en')
}
\author{
Angel Zazo, Department of Computer Science and Automatics, University of Salamanca
}
