% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{w_FilmInfo}
\alias{w_FilmInfo}
\title{Gets films' properties from Wikidata}
\usage{
w_FilmInfo(film_entity, langsorder = "en", wikilangs = "", format = "reduced")
}
\arguments{
\item{film_entity}{Wikidata entity to search for properties.}

\item{langsorder}{Order of languages in which the information will be
returned, separated with '|'. If no information is given in the first
language, next is used. This parameter is mandatory, at least one language is
required, default, 'en'.
Note: additional film title in Spanish is also returned.}

\item{wikilangs}{List of languages to limit the search of Wikipedia pages,
using "|" as separator. Wikipedias pages are returned in same order as
languages in this parameter. If wikilangs='' the function returns Wikipedia
pages in any language, not sorted.
#' @param format Wikipedia address format. By default is reduced, Otherwise, format is regular.}
}
\value{
: A data-frame with all properties of film_entity
}
\description{
Gets some properties of the Wikidata "film_entity" related to information about that film entity.
}
\examples{
\dontrun{
BenHur    <- w_FilmInfo(film_entity='Q180098', langsorder='es|en', wikilangs = 'es|fr')
Nosferatu <- w_FilmInfo(film_entity='Q151895', langsorder='es|en', wikilangs = 'es|fr|en')
# Nosferatu has public video:
Nosferatu$video
# Combination of information:
films <- rbind(BenHur, Nosferatu)
}
}
\author{
Angel Zazo, Department of Computer Science and Automatics, University of Salamanca
}
