% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{w_SearchByLabel_Exact}
\alias{w_SearchByLabel_Exact}
\title{w_SearchByLabel_Exact(string, langs='en', instanceof="", Pproperty="")
Search Wikidata entities by exact search using case sensitive and
differentiate diacritics in label and altLabel ("Also known as") in the
languages indicated in langs.}
\usage{
w_SearchByLabel_Exact(string, langs = "en", instanceof = "", Pproperty = "")
}
\arguments{
\item{string}{string (label or altLabel) to search.}

\item{langs}{Languages in which the information will be searched, using "|"
as separator. It also is the order of languages to obtain information.
This parameter is mandatory, at least one language is required, default, 'en'.}

\item{instanceof}{Wikidata entity of which the entities searched for are an
example or member of it (class). For example, if instanceof=Q5 the
search are filtered to Wikidata entities of class "Q5 = human". Some
entity classes are allowed, separated with '|'.}

\item{Pproperty}{Wikidata properties, separated with '|', to optionally
include in the search. For example, is Pproperty="P21", the results
include information of the sex of entities found as STRING.}
}
\value{
: A data-frame with 'entity', 'entityLabel', 'entityDescription' and,
additionally the properties of Pproperty.
}
\description{
w_SearchByLabel_Exact(string, langs='en', instanceof="", Pproperty="")
Search Wikidata entities by exact search using case sensitive and
differentiate diacritics in label and altLabel ("Also known as") in the
languages indicated in langs.
}
\examples{
df <- w_SearchByLabel_Exact(string='Iranzo', langs='es|en')
df <- w_SearchByLabel_Exact(string='Iranzo', langs='es|en', instanceof = 'Q5')
df <- w_SearchByLabel_Exact(string='Iranzo', langs='es|en', instanceof = 'Q5|Q101352')
df <- w_SearchByLabel_Exact(string='Iranzo', langs='es|en', instanceof = 'Q5',
Pproperty = 'P21|P569|P570')
}
\author{
Angel Zazo, Department of Computer Science and Automatics, University of Salamanca
}
