% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{w_SearchByLabel_Inlabel}
\alias{w_SearchByLabel_Inlabel}
\title{Search Wikidata entities matching whole words in any position}
\usage{
w_SearchByLabel_Inlabel(
  string,
  langsorder = "en",
  lang = "",
  instanceof = "",
  Pproperty = ""
)
}
\arguments{
\item{string}{string (label or altLabel) to search.}

\item{langsorder}{Order of languages in which the information will be
returned, separated with '|'. If no information is given in the first
language, next is used. This parameter is mandatory, at least one language is
required, default, 'en'.}

\item{lang}{The language to search, only one. If lang="", search is in any
language.}

\item{instanceof}{Wikidata entity of which the entities searched for are an
example or member of it (class). For example, if instanceof=Q5 the
search are filtered to Wikidata entities of class Q5 (human). Some
entity classes are allowed, separated with '|'.}

\item{Pproperty}{Wikidata properties, separated with '|', to optionally
include in the search. For example, is Pproperty="P21", the results
include information of the sex of entities found as STRING.}
}
\value{
: A data-frame with 'entity', 'entityLabel', 'entityDescription' and,
additionally the properties of Pproperty.
}
\description{
Search Wikidata entities matching whole words in any position in label and
altLabel. Diacritics and case are ignored. If lang has a value (es, en...)
then the search is only in that language, otherwise any.
}
\examples{
## Search in any position in Label or AltLabel (diacritics and case are ignored)
# If lang=='' search in any language, else the search is performed only in the
# language indicated.
#   w_SearchByLabel_Inlabel(string, langsorder, lang, instanceof, Pproperty)
df <- w_SearchByLabel_Inlabel(string='Iranzo', langsorder='es|en')
# Search in Chinese (Simplified) (language code: zh):
df <- w_SearchByLabel_Inlabel(string='Iranzo', langsorder='zh|es', lang='zh')
# Optional instanceof and Pproperty
df <- w_SearchByLabel_Inlabel(string='Iranzo', langsorder='es|en', instanceof = 'Q5')
df <- w_SearchByLabel_Inlabel(string='Iranzo', langsorder='es|en', instanceof = 'Q5|Q101352')
df <- w_SearchByLabel_Inlabel(string='Iranzo', langsorder='es|en', instanceof = 'Q5',
Pproperty = 'P21|P569|P570')
}
\author{
Angel Zazo, Department of Computer Science and Automatics, University of Salamanca
}
