% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{WDQS_EntityInfo_tiny}
\alias{WDQS_EntityInfo_tiny}
\title{WDQS_EntityInfo_tiny(entity, langs, wikilangs)
Same as WDQS_EntityInfo function but less properties are requested and
less checks are done.
Gets some properties of the Wikidata "entity" related to birth and death
dates, places, occupations, works, education, awards, identificator in some
libraries, wikipedia pages (only languages in wikilangs parameter), etc.}
\usage{
WDQS_EntityInfo_tiny(
  entity,
  langs = "es|en|ca|eu|gl|pt|fr|it|de|ru|ko|ja|zh",
  wikilangs = "es|en|ca|eu|gl|pt|fr|it|de|ru|ko|ja|zh"
)
}
\arguments{
\item{entity}{Wikidata entity to search for properties.}

\item{langs}{Language order of the "SERVICE wikibase:label", mandatory.
Note: sometimes not label in any language of langs is assigned to a entity,
so an additional search is used to obtain almost one label for it
(?entitylab) with LIMIT 1.}

\item{wikilangs}{only return the Wikipedia pages of the entity in those
languages and in its order. If wikilangs="" retrieve all wikipedias pages,
not sorted.}
}
\value{
: A data-frame with all properties of entity
}
\description{
WDQS_EntityInfo_tiny(entity, langs, wikilangs)
Same as WDQS_EntityInfo function but less properties are requested and
less checks are done.
Gets some properties of the Wikidata "entity" related to birth and death
dates, places, occupations, works, education, awards, identificator in some
libraries, wikipedia pages (only languages in wikilangs parameter), etc.
}
\author{
Angel Zazo, Department of Computer Science and Automatics, University of Salamanca
}
