% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{WDQS_FilmInfo}
\alias{WDQS_FilmInfo}
\title{WDQS_FilmInfo(film_entity, langs, wikilangs)
Gets some properties of the Wikidata "film_entity" related to information
about that film entity.}
\usage{
WDQS_FilmInfo(
  film_entity,
  langs = "es|en|ca|eu|gl|pt|fr|it|de|ru|ko|ja|zh",
  wikilangs = "es|en|ca|eu|gl|pt|fr|it|de|ru|ko|ja|zh"
)
}
\arguments{
\item{film_entity}{Wikidata entity to search for properties.}

\item{langs}{Language order of the "SERVICE wikibase:label", mandatory.
Note: sometimes not label in any language of langs is assigned to a entity,
so an additional search is used to obtain almost one label for it
(?entitylab) with LIMIT 1.}

\item{wikilangs}{only return the Wikipedia pages of the entity in those
languages and in its order. If wikilangs="" retrieve all wikipedias pages,
not sorted.}
}
\value{
: A data-frame with all properties of film_entity
}
\description{
WDQS_FilmInfo(film_entity, langs, wikilangs)
Gets some properties of the Wikidata "film_entity" related to information
about that film entity.
}
\author{
Angel Zazo, Department of Computer Science and Automatics, University of Salamanca
}
