% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{WDQS_isInstanceOf}
\alias{WDQS_isInstanceOf}
\title{WDQS_isInstanceOf(entity_list, instanceof)
Search using WDQS if the Wikidata entities in entity_list are instances of
"instanceof" Wikipedia class. For example, if instanceof="Q5", search if
entities are instances of the Wikidata class Q5, i.e, are humans.
Duplicated entities are deleted before search.}
\usage{
WDQS_isInstanceOf(entity_list, instanceof)
}
\arguments{
\item{entity_list}{A vector with de Wikidata entities.}

\item{instanceof}{The Wikidata class to check, madatory.}
}
\value{
A data-frame with two columns, first Wikidata entity, second True
if that entity is instance of the "instanceof" entity, else False. Index of
data-frame are also set to entity_list.
}
\description{
WDQS_isInstanceOf(entity_list, instanceof)
Search using WDQS if the Wikidata entities in entity_list are instances of
"instanceof" Wikipedia class. For example, if instanceof="Q5", search if
entities are instances of the Wikidata class Q5, i.e, are humans.
Duplicated entities are deleted before search.
}
\author{
Angel Zazo, Department of Computer Science and Automatics, University of Salamanca
}
