% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{WDQS_isValid}
\alias{WDQS_isValid}
\title{WDQS_isValid(entity_list, nlimit=50000)
Check if the Wikidata entities are valid. A entity is valid if it has a label
or has a description. If one entity exists but is not valid, is possible that
it has a redirection to other entity, in that case, the redirection is
obtained. Other entities may have existed in the past, but have been deleted.
Duplicated entities in entity_list are deleted before checking. Index of the
data-frame returned are also set to entity_list.}
\usage{
WDQS_isValid(entity_list, nlimit = 50000)
}
\arguments{
\item{entity_list}{A vector with de Wikidata entities.}

\item{nlimit}{If the number of entities exceeds this number, do chunked
queries. This is the number of entities requested in each chunk.}
}
\value{
A data-frame with three columns: firts, the entity itself, second,
if that entity is valid in Wikidata (TRUE or FALSE), last, if the entity
redirects to another Wikidata entity, this entity.
}
\description{
WDQS_isValid(entity_list, nlimit=50000)
Check if the Wikidata entities are valid. A entity is valid if it has a label
or has a description. If one entity exists but is not valid, is possible that
it has a redirection to other entity, in that case, the redirection is
obtained. Other entities may have existed in the past, but have been deleted.
Duplicated entities in entity_list are deleted before checking. Index of the
data-frame returned are also set to entity_list.
}
\examples{
\dontrun{
WDQS_isValid(c("Q9021", "Q115637688", "Q105660123"))

l  <- WDQS_OccupationEntities(Qoc='Q2306091')
l2 <- append(l, c("Q115637688", "Q105660123"))  # Note: adding two new entities
v <- WDQS_isValid(l2)
# Not valid
v[!v$valid,]
}
}
\author{
Angel Zazo, Department of Computer Science and Automatics, University of Salamanca
}
