% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{WDQS_OccupationCount}
\alias{WDQS_OccupationCount}
\title{WDQS_OccupationCount(Qoc)
Search in WDQS to know the number of Wikidata entities with P106 property
(occupation) set to Qoc.}
\usage{
WDQS_OccupationCount(Qoc)
}
\arguments{
\item{Qoc}{The Wikidata entity of the occupation. For example, Q2306091
sociologist, Q2526255 Film director, etc.}
}
\value{
The number of entities with that occupation (integer).
}
\description{
WDQS_OccupationCount(Qoc)
Search in WDQS to know the number of Wikidata entities with P106 property
(occupation) set to Qoc.
}
\examples{
WDQS_OccupationCount(Qoc='Q2306091') # Qoc for Sociologist
}
\author{
Angel Zazo, Department of Computer Science and Automatics, University of Salamanca
}
