% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{WDQS_OccupationEntities}
\alias{WDQS_OccupationEntities}
\title{WDQS_OccupationEntities(Qoc, nlimit)
Returns the Wikidata entities of the occupation Qoc. Use chunked requests.}
\usage{
WDQS_OccupationEntities(Qoc, nlimit = 10000)
}
\arguments{
\item{Qoc}{The Wikidata entity of the occupation. For example, Q2306091 (sociologist)}

\item{nlimit}{If the number of entities found with that occupation exceeds
this number, do chunked queries. This is the number of entities requested
in each chunk.}
}
\value{
A vector with the Wikidata entities with that occupation.
}
\description{
WDQS_OccupationEntities(Qoc, nlimit)
Returns the Wikidata entities of the occupation Qoc. Use chunked requests.
}
\author{
Angel Zazo, Department of Computer Science and Automatics, University of Salamanca
}
