% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{WDQS_OccupationEntitiesyWikipedias}
\alias{WDQS_OccupationEntitiesyWikipedias}
\title{WDQS_OccupationEntitiesyWikipedias(Qoc, wikilangs, nlimit)
Returns the Wikidata entities of the occupation Qoc, plus the Wikipedia pages
of them. Return all Wikipedias pages, not limited by languages. Use chunked
requests. Use chunked requests.
Note that WDQS_OccupationEntityWikipedias is similar to first launch
WDQS_OccupationEntity and then launch WDQS_Wikipedias, but is more efficient.}
\usage{
WDQS_OccupationEntitiesyWikipedias(Qoc, nlimit = 5000)
}
\arguments{
\item{Qoc}{The Wikidata entity of the occupation. For example, Q2306091
(sociologist)}

\item{nlimit}{If the number of entities found with that occupation exceeds
this number, do chunked queries. This is the number of entities requested
in each chunk.}
}
\value{
A data-frame with Wikidata entities and Wikipedia pages concatenated with "|".
}
\description{
WDQS_OccupationEntitiesyWikipedias(Qoc, wikilangs, nlimit)
Returns the Wikidata entities of the occupation Qoc, plus the Wikipedia pages
of them. Return all Wikipedias pages, not limited by languages. Use chunked
requests. Use chunked requests.
Note that WDQS_OccupationEntityWikipedias is similar to first launch
WDQS_OccupationEntity and then launch WDQS_Wikipedias, but is more efficient.
}
\author{
Angel Zazo, Department of Computer Science and Automatics, University of Salamanca
}
