% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{WDQS_Property}
\alias{WDQS_Property}
\title{WDQS_Property(entity_list, Pproperty, langsorder)
Search the entities of the entity_list for property or properties. Return the
properties in langsorder order. Duplicated entities are deleted before search.}
\usage{
WDQS_Property(
  entity_list,
  Pproperty,
  langsorder = "es|en|ca|eu|gl|pt|fr|it|de|ru|ko|ja|zh"
)
}
\arguments{
\item{entity_list}{A vector with de Wikidata entities.}

\item{Pproperty}{Wikidata properties to search, separated with '|'. For
example, is Pproperty="P21", the results contain information of the sex
of entities. If Pproperty="P21|P569" also searches for birthdate. If
Pproperty='P21|P569|P214' also searches for VIAF identifier.}

\item{langsorder}{Order of languages in which the information will be returned.}
}
\value{
: A data-frame with 'entity', 'entityLabel', 'entityDescription' and,
additionally, the properties of Pproperty. Index of the data-frame is also
set to entity_list.
}
\description{
WDQS_Property(entity_list, Pproperty, langsorder)
Search the entities of the entity_list for property or properties. Return the
properties in langsorder order. Duplicated entities are deleted before search.
}
\author{
Angel Zazo, Department of Computer Science and Automatics, University of Salamanca
}
