% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{WDQS_Redirection}
\alias{WDQS_Redirection}
\title{WDQS_Redirection(entity_list)
Returns the Wikidata entities which are the redirection of the entities in
entity_list. Return a data-frame with entities and values the redirection
(destiny) or NA if not exists the destiny or entity not exits or it is
deleted. Duplicated entities are deleted before search.}
\usage{
WDQS_Redirection(entity_list)
}
\arguments{
\item{entity_list}{A vector with de Wikidata entities.}
}
\value{
: A data-frame with properties of entity about redirections. Index of
data-frame are also set to entity_list.
WDQS_Redirection(c("Q9021", "Q115637688", "Q105660123"))
Q9021       -> NA [Not redirect]
Q115637688  -> NA [Was Deleted]
Q105660123  -> Q97352588 [Redirect to Q97352588]
}
\description{
WDQS_Redirection(entity_list)
Returns the Wikidata entities which are the redirection of the entities in
entity_list. Return a data-frame with entities and values the redirection
(destiny) or NA if not exists the destiny or entity not exits or it is
deleted. Duplicated entities are deleted before search.
}
\author{
Angel Zazo, Department of Computer Science and Automatics, University of Salamanca
}
