% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{WDQS_SearchByAuthority}
\alias{WDQS_SearchByAuthority}
\title{WDQS_SearchByAuthority(authority, instanceof, langsorder)
Search for Wikidata entities that have an identifier in the Wikidata
authority class "authority". Return the properties in the language order
indicated in langsorder. If instanceof has a value, then response is limited
to entities which are instance of it.}
\usage{
WDQS_SearchByAuthority(
  authority,
  instanceof = "",
  langsorder = "es|en|ca|eu|gl|pt|fr|it|de|ru|ko|ja|zh"
)
}
\arguments{
\item{authority}{Wikidata identifier class for authority. For example, is
authority = "P4439", then search Wikidata entities that have an identifier
in the MNCARS (Museo Nacional Centro de Arte Reina Sofía) database.}

\item{instanceof}{Wikidata entity of which the entities searched for are an
example or member of it (class). For example, if instanceof=Q5 the
search are filtered to Wikidata entities of class "Q5 = human".}

\item{langsorder}{Order of languages in which the information will be returned.}
}
\value{
: A data-frame with 'entity', 'entityLabel', 'entityDescription' and,
additionally the properties of Pproperty.
}
\description{
WDQS_SearchByAuthority(authority, instanceof, langsorder)
Search for Wikidata entities that have an identifier in the Wikidata
authority class "authority". Return the properties in the language order
indicated in langsorder. If instanceof has a value, then response is limited
to entities which are instance of it.
}
\author{
Angel Zazo, Department of Computer Science and Automatics, University of Salamanca
}
