% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{WDQS_SearchByLabel_Inlabel}
\alias{WDQS_SearchByLabel_Inlabel}
\title{WDQS_SearchByLabel_Inlabel(string, instanceof, Pproperty, lang, langsorder)
Search Wikidata entities matching whole words in any position in label and
altLabel. Diacritics and case are ignored. If lang has a value (es, en...)
then the search is only in that language, otherwise any.}
\usage{
WDQS_SearchByLabel_Inlabel(
  string,
  instanceof = "",
  Pproperty = "",
  lang = "",
  langsorder = "es|en|ca|eu|gl|pt|fr|it|de|ru|ko|ja|zh"
)
}
\arguments{
\item{string}{string (label or altLabel) to search.}

\item{instanceof}{Wikidata entity of which the entities searched for are an
example or member of it (class). For example, if instanceof=Q5 the
search are filtered to Wikidata entities of class "Q5 = human".}

\item{Pproperty}{Wikidata properties, separated with '|', to optionally
include in the search. For example, is Pproperty="P21", the results
include information of the sex of entities found as STRING.}

\item{lang}{the language to search. If lang="", search is in any language.}

\item{langsorder}{Order of languages in which the information will be returned.}
}
\value{
: A data-frame with 'entity', 'entityLabel', 'entityDescription' and,
additionally the properties of Pproperty.
}
\description{
WDQS_SearchByLabel_Inlabel(string, instanceof, Pproperty, lang, langsorder)
Search Wikidata entities matching whole words in any position in label and
altLabel. Diacritics and case are ignored. If lang has a value (es, en...)
then the search is only in that language, otherwise any.
}
\author{
Angel Zazo, Department of Computer Science and Automatics, University of Salamanca
}
