% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{WDQS_SearchByLabel_Startswith}
\alias{WDQS_SearchByLabel_Startswith}
\title{WDQS_SearchByLabel_Startswith(string, instanceof, Pproperty, lang, langsorder)
Search Wikidata entities which label or altLabel starts with "string" in
language "lang". It's similar to a wildcard search: "string*". Diacritics and
case are ignored. Search "string" in language "lang" (mandatory) in label,
but in any language in altLabel.}
\usage{
WDQS_SearchByLabel_Startswith(
  string,
  instanceof = "",
  Pproperty = "",
  lang = "es",
  langsorder = "es|en|ca|eu|gl|pt|fr|it|de|ru|ko|ja|zh"
)
}
\arguments{
\item{string}{string to search.}

\item{instanceof}{Wikidata entity of which the entities searched for are an
example or member of it (class). For example, if instanceof=Q5 the
search are filtered to Wikidata entities of class Q5 (human).}

\item{Pproperty}{Wikidata properties, separated with '|', to optionally
include in the search. For example, is Pproperty="P21", the results
include information of the sex of entities found.}

\item{lang}{language to search in label, mandatory.}

\item{langsorder}{Order of languages in which the information will be returned.}
}
\value{
: A data-frame with 'entity', 'entityLabel', 'entityDescription',
'instance', 'instanceLabel', 'altLabel' and, additionally, the properties of
Pproperty param.
}
\description{
WDQS_SearchByLabel_Startswith(string, instanceof, Pproperty, lang, langsorder)
Search Wikidata entities which label or altLabel starts with "string" in
language "lang". It's similar to a wildcard search: "string*". Diacritics and
case are ignored. Search "string" in language "lang" (mandatory) in label,
but in any language in altLabel.
}
\author{
Angel Zazo, Department of Computer Science and Automatics, University of Salamanca
}
