% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{Wikidata_occupationCount}
\alias{Wikidata_occupationCount}
\title{Wikidata_occupationCount
Search Wikidata Query Service (WDQS) to know the number of Wikidata entities with P106 property (occupation) set to Qoc.}
\usage{
Wikidata_occupationCount(Qoc = "")
}
\arguments{
\item{Qoc}{The Wikidata entity of the occupation}
}
\value{
The number of entities with that occupation (integer)
}
\description{
Wikidata_occupationCount
Search Wikidata Query Service (WDQS) to know the number of Wikidata entities with P106 property (occupation) set to Qoc.
}
\examples{
Wikidata_occupationCount('Q2526255') # Film director
}
\author{
Angel F. Zazo, Departament of Computer Science and Automatics, University of Salamanca
}
