% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{Wikidata_Wikipedias}
\alias{Wikidata_Wikipedias}
\title{Wikidata_wikipedias
For an occupation, obtains all Wikidata entity of the people with that ocupation, also the
number of wikipedias witch article they have, and the URL of those wikipedias (sep = tab)
Run queries splitting with offset and chunk.}
\usage{
Wikidata_Wikipedias(Qoc, chunk = 10000)
}
\arguments{
\item{Qoc}{The Wikidata entity of the occupation}

\item{chunk}{The chunk to split intermediate results with the aim of reduce the limit 60 seconds processig time.}
}
\value{
A list with Wikidata entity, the number of wikipedias and the URL
}
\description{
Wikidata_wikipedias
For an occupation, obtains all Wikidata entity of the people with that ocupation, also the
number of wikipedias witch article they have, and the URL of those wikipedias (sep = tab)
Run queries splitting with offset and chunk.
}
\author{
Angel F. Zazo, Departament of Computer Science and Automatics, University of Salamanca
}
