% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{Wikimedia_get_redirects}
\alias{Wikimedia_get_redirects}
\title{Wikimedia_get_redirects
Obtains redirection pages (from namespace 0) to the article page in the Wikimedia project}
\usage{
Wikimedia_get_redirects(article, project = "en.wikipedia.org")
}
\arguments{
\item{article}{Article target}

\item{project}{Wikimedia project, defaults "en.wikipedio.org"}
}
\value{
A list with the firts element the target of all redirections, or NULL on error.
}
\description{
Wikimedia_get_redirects
Obtains redirection pages (from namespace 0) to the article page in the Wikimedia project
}
\author{
Angel F. Zazo, Departament of Computer Science and Automatics, University of Salamanca
}
