% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{Wikimedia_query}
\alias{Wikimedia_query}
\title{Wikimedia_query
Use httr package to retrieve responses in JSON format about an article from Wikimedia API.}
\usage{
Wikimedia_query(
  query,
  project = "en.wikipedia.org",
  headers = my_headers,
  attempts = 2
)
}
\arguments{
\item{query}{A list with de (key, values) pairs with the search.}

\item{project}{The Wikimedia project to search.}

\item{headers}{A vector with aditional query headers for the request.}

\item{attempts}{On errors, the maximun number of times the query is launched if repetition_on_error is not zero (default 2)}
}
\value{
The response in JSON format or NULL on errors.
}
\description{
Wikimedia_query
Use httr package to retrieve responses in JSON format about an article from Wikimedia API.
}
\author{
Angel F. Zazo, Departament of Computer Science and Automatics, University of Salamanca
}
