% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{WM_Pageviews}
\alias{WM_Pageviews}
\title{WM_Pageviews
Use "Wikimedia REST API (https://wikimedia.org/api/rest_v1/) to search the
number of views one article has in a Wikimedia project in a date interval
(see granularity). Optionally include redirections to the article page.}
\usage{
WM_Pageviews(
  article,
  start,
  end,
  project = "en.wikipedia.org",
  access = "all-access",
  agent = "user",
  granularity = "monthly",
  include_redirects = FALSE
)
}
\arguments{
\item{article}{The title of the article to search}

\item{start, end}{First and last day to include (format YYYYMMDD or YYYYMMDDHH)}

\item{project}{The Wikimedia project, defaults en.wikipedia.org}

\item{access}{Filter by access method: all-access (default), desktop, mobile-app, mobile-web}

\item{agent}{Filter by agent type: all-agents, user (default), spider, automated}

\item{granularity}{Time unit for the response data: daily, monthly (default)}

\item{include_redirects}{Boolean to include redirection to the article page (defaults: False)}
}
\value{
A vector with the number of visits by granularity
}
\description{
WM_Pageviews
Use "Wikimedia REST API (https://wikimedia.org/api/rest_v1/) to search the
number of views one article has in a Wikimedia project in a date interval
(see granularity). Optionally include redirections to the article page.
}
\author{
Angel Zazo, Department of Computer Science and Automatics, University of Salamanca
}
