% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{wmflabs_get_allinfo}
\alias{wmflabs_get_allinfo}
\title{wmflabs_get_allinfo
obtains info about infotype = (articleinfo | prose | links) about an article in the
wikimedia project}
\usage{
wmflabs_get_allinfo(
  article,
  project = "en.wikipedia.org",
  links_in_count_redirects = FALSE
)
}
\description{
wmflabs_get_allinfo
obtains info about infotype = (articleinfo | prose | links) about an article in the
wikimedia project
}
