% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipedia-tools.R
\name{add_wikipedia_matches}
\alias{add_wikipedia_matches}
\title{Search Wikipedia and add match columns to a data frame}
\usage{
add_wikipedia_matches(
  df,
  name_col = "name",
  lang = "en",
  delay = 0.5,
  limit = 5
)
}
\description{
Search Wikipedia and add match columns to a data frame
}
