% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipedia-tools.R
\name{cache_wikitext}
\alias{cache_wikitext}
\title{Cache wikitext to a local file; load from cache if available}
\usage{
cache_wikitext(title, cache_dir, lang = "en")
}
\arguments{
\item{title}{Page title}

\item{cache_dir}{Directory path for cached files}

\item{lang}{Language code}
}
\value{
Wikitext string (NULL if page doesn't exist)
}
\description{
Cache wikitext to a local file; load from cache if available
}
