% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{checkTitles}
\alias{checkTitles}
\title{checkTitles(titles)
Check if titles are valid. Return TRUE is all titles are valid, else FALSE.
See:See https://en.wikipedia.org/wiki/Wikipedia:Page_name#Technical_restrictions_and_limitations}
\usage{
checkTitles(titles)
}
\arguments{
\item{titles}{A vector of titles to check.}
}
\description{
checkTitles(titles)
Check if titles are valid. Return TRUE is all titles are valid, else FALSE.
See:See https://en.wikipedia.org/wiki/Wikipedia:Page_name#Technical_restrictions_and_limitations
}
