% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipedia-tools.R
\name{count_refs}
\alias{count_refs}
\title{Count Reference Tags in Wikitext}
\usage{
count_refs(wikitext)
}
\arguments{
\item{wikitext}{Character. Raw wikitext string.}
}
\value{
Integer. Number of `<ref>` tags found. Returns `0L` if `wikitext`
  is `NULL`.
}
\description{
Counts `<ref>` opening tags in wikitext, including self-closing tags
(`<ref name="x" />`). Counts distinct inline references, not closing
`</ref>` tags. For citation template counts, see [count_citations()].
}
\examples{
wt <- "<ref>Smith 2000</ref> text <ref name='jones'/> more"
count_refs(wt)  # 2

}
\seealso{
[count_citations()]
}
