% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipedia-tools.R
\name{get_wikitable}
\alias{get_wikitable}
\title{Format a data frame as a MediaWiki wikitable string}
\usage{
get_wikitable(
  df,
  caption = NULL,
  class = "wikitable sortable",
  column_names = NULL
)
}
\description{
Format a data frame as a MediaWiki wikitable string
}
