% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipedia-tools.R
\name{get_wikitext_by_name}
\alias{get_wikitext_by_name}
\title{Fetch raw wikitext for a named article}
\usage{
get_wikitext_by_name(article_name, lang = "en")
}
\description{
Fetch raw wikitext for a named article
}
