% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipedia-tools.R
\name{get_wikitext_by_revid}
\alias{get_wikitext_by_revid}
\title{Fetch raw wikitext for a specific revision}
\usage{
get_wikitext_by_revid(article_name, revision_id, lang = "en")
}
\description{
Fetch raw wikitext for a specific revision
}
