% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{m_Redirects}
\alias{m_Redirects}
\title{m_Redirects(titles, project="en.wikipedia.org")
Obtains the redirection pages to the article titles in the Wikimedia project
restricted to namespace 0. Return a vector for each title, in each vector the
first element is the page destiny, rest are all pages that redirect to it. If
a title is not in the Wikimedia project its list is NA.}
\usage{
m_Redirects(titles, project = "en.wikipedia.org")
}
\arguments{
\item{titles}{A vector of page titles to search for.}

\item{project}{Wikimedia project, defaults "en.wikipedia.org"}
}
\value{
A vector for each title, with all pages that are redirects to the
first element.
}
\description{
m_Redirects(titles, project="en.wikipedia.org")
Obtains the redirection pages to the article titles in the Wikimedia project
restricted to namespace 0. Return a vector for each title, in each vector the
first element is the page destiny, rest are all pages that redirect to it. If
a title is not in the Wikimedia project its list is NA.
}
\examples{
a <- m_Redirects(c('Cervantes', 'Planck', 'Noexiste'), project='es.wikipedia.org')
}
\author{
Angel Zazo, Department of Computer Science and Automatics, University of Salamanca
}
