% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{MW_XtoolsInfoAll}
\alias{MW_XtoolsInfoAll}
\title{MW_XtoolsInfoAll
Obtains information about an article in the Wikimedia project using wmflabs
in JSON format, or NULL on error. The XTools Page API endpoints offer data
related to a single page. See https://www.mediawiki.org/wiki/XTools/API/Page
Obtains all information (articleinfo, prose and links) about an article in
the wikimedia project.}
\usage{
MW_XtoolsInfoAll(
  article,
  project = "en.wikipedia.org",
  inlinks_redirects = FALSE
)
}
\arguments{
\item{article}{The title of the article to search}

\item{project}{The Wikimedia project, defaults en.wikipedia.org}

\item{inlinks_redirects}{If infotype==links, if redirects are included or not}
}
\value{
The information about the article.
Note: use the MW_XtoolsInfo function.
}
\description{
MW_XtoolsInfoAll
Obtains information about an article in the Wikimedia project using wmflabs
in JSON format, or NULL on error. The XTools Page API endpoints offer data
related to a single page. See https://www.mediawiki.org/wiki/XTools/API/Page
Obtains all information (articleinfo, prose and links) about an article in
the wikimedia project.
}
