% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_quick_statement.R
\name{remove_quick_statement_column}
\alias{remove_quick_statement_column}
\title{Add a QuickStatements Column for Removing Statements to a Data Frame}
\usage{
remove_quick_statement_column(dataframe, qid_col, property, value_col, ...)
}
\arguments{
\item{dataframe}{A data frame (or tibble).}

\item{qid_col}{Unquoted column name containing Wikidata item IDs.}

\item{property}{Character. The property ID (e.g., \code{"P123"}).}

\item{value_col}{Unquoted column name containing statement values.}

\item{...}{Additional arguments passed to `create_quick_statement()`.}
}
\value{
The input data frame with an additional \code{quick_statement} column
 containing negated statements.
}
\description{
Like \code{add_quick_statement_column} but negates the statement by prepending a
"-" to the command. This is useful for removing statements via QuickStatements.
Note that qualifiers and references cannot be used with negated statements.
The \code{property} argument must be a statement property (e.g., \code{"P123"}).
}
\examples{
departments <- tibble::tribble(
~qid,             ~label_en, ~cod.dep,
"Q233169",     "Beni Department",     "08",
"Q233917", "Cochabamba Department",     "03",
"Q233933",   "Tarija Department",     "06",
"Q235106", "Santa Cruz Department",     "07",
"Q235110", "Chuquisaca Department",     "01",
"Q235362",    "Pando Department",     "09",
"Q238079",   "Potosí Department",     "05",
"Q232784",    "La Paz Department",     "02",
"Q844510",             "Litoral",       NA,
"Q1061368",    "Oruro Department",     "04"
)
departments \%>\%
  remove_quick_statement_column(qid, "P14142", cod.dep)

}
