% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipedia-tools.R
\name{split_on_top_level_pipes}
\alias{split_on_top_level_pipes}
\title{Split a string on "|" characters that are not inside {{ }}}
\usage{
split_on_top_level_pipes(text)
}
\description{
Split a string on "|" characters that are not inside {{ }}
}
