% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\docType{data}
\name{user_agent}
\alias{user_agent}
\title{See https://meta.wikimedia.org/wiki/User-Agent_policy.}
\format{
An object of class \code{character} of length 1.
}
\usage{
user_agent
}
\description{
See https://meta.wikimedia.org/wiki/User-Agent_policy.
}
\keyword{datasets}
