% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{v_Sources}
\alias{v_Sources}
\title{Gets text of all sources from a VIAF record}
\usage{
v_Sources(viaf)
}
\arguments{
\item{viaf}{VIAF cluster record (in JSON format).}
}
\value{
A data-frame with text and sources.
}
\description{
Return the text of all sources id from the VIAF record. Note that the VIAF
record musts be in JSON format.
}
