% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{v_Titles}
\alias{v_Titles}
\title{Gets works titles from a VIAF record}
\usage{
v_Titles(viaf)
}
\arguments{
\item{viaf}{VIAF cluster record (in JSON format).}
}
\value{
A list with titles.
}
\description{
Returns titles of works from the VIAF record. Note that the VIAF record musts
be in JSON format.
}
