% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{VIAF_Search}
\alias{VIAF_Search}
\title{VIAF_Search(CQL_Query)
Run the CQL_Query using the VIAF Search API and return a list of records
found. The search string is formed using the CQL_Query syntax of the API.
Note that returned records use the http://viaf.org/BriefVIAFCluster record
schema. Also note that only returns a maximum of 250 records.}
\usage{
VIAF_Search(CQL_Query)
}
\arguments{
\item{CQL_Query}{string with the search. See https://www.oclc.org/developer/api/oclc-apis/viaf/authority-cluster.en.html}
}
\value{
A list with the records found.
}
\description{
VIAF_Search(CQL_Query)
Run the CQL_Query using the VIAF Search API and return a list of records
found. The search string is formed using the CQL_Query syntax of the API.
Note that returned records use the http://viaf.org/BriefVIAFCluster record
schema. Also note that only returns a maximum of 250 records.
}
