% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{w_OccupationEntities}
\alias{w_OccupationEntities}
\title{w_OccupationEntities(Qoc, nlimit=10000)
Returns the Wikidata entities which have the occupation indicated in Qoc, the
Wikidata entity for that occupation. Use chunked requests.}
\usage{
w_OccupationEntities(Qoc, nlimit = 10000)
}
\arguments{
\item{Qoc}{The Wikidata entity of the occupation. For example, Q2306091 (sociologist)}

\item{nlimit}{If the number of entities found with that occupation exceeds
this number, chunked queries are done. This is the number of entities requested
in each chunk.}
}
\value{
A vector with the Wikidata entities with that occupation.
}
\description{
w_OccupationEntities(Qoc, nlimit=10000)
Returns the Wikidata entities which have the occupation indicated in Qoc, the
Wikidata entity for that occupation. Use chunked requests.
}
\examples{
l <- w_OccupationEntities(Qoc='Q2306091')
}
\author{
Angel Zazo, Department of Computer Science and Automatics, University of Salamanca
}
