% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{w_OccupationEntitiesWikipedias}
\alias{w_OccupationEntitiesWikipedias}
\title{Gets Wikipedias entries with a certain occupation}
\usage{
w_OccupationEntitiesWikipedias(Qoc, nlimit = 5000)
}
\arguments{
\item{Qoc}{The Wikidata entity of the occupation. For example, Q2306091 is
the entity of sociologist.}

\item{nlimit}{If the number of entities found with that occupation exceeds
this number, chunked queries are done. This is the number of entities requested
in each chunk.}
}
\value{
A data-frame with Wikidata entities, the number of Wikipedias in
which they have page, the Wikipedia languages, the page titles, and,
finally, the URL to the pages. Last three columns concatenated with "|".
}
\description{
Returns the Wikidata entities which have the occupation indicated in Qoc,
plus the Wikipedia page titles of them. Return all Wikipedias pages, not
limited by languages. Use chunked requests.
Note that w_OccupationEntitiesWikipedias is similar to first launch
w_OccupationEntities and then launch w_Wikipedias, but is more efficient.
}
\examples{
\dontrun{
lw <- w_OccupationEntitiesWikipedias(Qoc='Q2306091')
}
}
\author{
Angel Zazo, Department of Computer Science and Automatics, University of Salamanca
}
