% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{WDQS_query}
\alias{WDQS_query}
\title{WDQS_query
Retrieves responses from Wikidata Query Service (WDQS). Uses ratelimitr if
param limit_requester = TRUE.}
\usage{
WDQS_query(
  sparql_query,
  format = "csv",
  method = "GET",
  limit_requester = FALSE
)
}
\arguments{
\item{sparql_query}{A string with the query in SPARQL language.}

\item{format}{A string with the query response format.
See https://www.mediawiki.org/wiki/Wikidata_Query_Service/User_Manual#SPARQL_endpoint.
Only  'json', 'xml' or 'csv' formats are permitted, default 'csv'.}

\item{method}{The method used in the httr request. Default 'GET'.}

\item{limit_requester}{If True, uses ratelimitr to limit the requests.}
}
\value{
The response in selected format or NULL on errors.
}
\description{
WDQS_query
Retrieves responses from Wikidata Query Service (WDQS). Uses ratelimitr if
param limit_requester = TRUE.
}
\author{
Angel Zazo, Department of Computer Science and Automatics, University of Salamanca
}
