% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{WDQS_SearchByLabel_Exact}
\alias{WDQS_SearchByLabel_Exact}
\title{WDQS_SearchByLabel_Exact(string, instanceof, Pproperty, langs)
Search Wikidata entities by exact search using case sensitive and
diferentiate diacritics in label and altLabel ("Also known as") in the
languages indicated in langs.}
\usage{
WDQS_SearchByLabel_Exact(
  string,
  instanceof = "",
  Pproperty = "",
  langs = "es|en|ca|eu|gl|pt|fr|it|de|ru|ko|ja|zh"
)
}
\arguments{
\item{string}{string (label or altLabel) to search.}

\item{instanceof}{Wikidata entity of which the entities searched for are an
example or member of it (class). For example, if instanceof=Q5 the
search are filtered to Wikidata entities of class "Q5 = human".}

\item{Pproperty}{Wikidata properties, separated with '|', to optionally
include in the search. For example, is Pproperty="P21", the results
include information of the sex of entities found as STRING.}

\item{langs}{Languages in which the information will be searched. It also
is the order of languages to obtain information.}
}
\value{
: A data-frame with 'entity', 'entityLabel', 'entityDescription' and,
additionally the properties of Pproperty.
}
\description{
WDQS_SearchByLabel_Exact(string, instanceof, Pproperty, langs)
Search Wikidata entities by exact search using case sensitive and
diferentiate diacritics in label and altLabel ("Also known as") in the
languages indicated in langs.
}
\author{
Angel Zazo, Department of Computer Science and Automatics, University of Salamanca
}
