% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wikidata_instances.R,
%   R/wikitools_package.R
\docType{package}
\name{wikitools-package}
\alias{wikitools-package}
\title{wikitools: Tools to Engage with Wikipedia and Wikidata}
\description{
Tools for retrieving, parsing, and matching Wikipedia and Wikidata content, including helpers for wikitext analysis and QuickStatements generation workflows.

Provides functions for retrieving, parsing, and matching Wikipedia and
Wikidata content, including helpers for wikitext analysis and
QuickStatements generation workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CarwilB/wikitools}
  \item \url{https://carwilb.github.io/wikitools/}
  \item Report bugs at \url{https://github.com/CarwilB/wikitools/issues}
}

}
\author{
\strong{Maintainer}: Carwil Bjork-James \email{c.bjork-james@vanderbilt.edu} (\href{https://orcid.org/0000-0002-4977-0458}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Carwil Bjork-James \email{c.bjork-james@vanderbilt.edu} (\href{https://orcid.org/0000-0002-4977-0458}{ORCID}) [copyright holder]
}

}
\keyword{internal}
