% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_utils.R
\name{wmflabs_get_allinfo}
\alias{wmflabs_get_allinfo}
\title{wmflabs_get_allinfo
Obtains information about an article in the Wikimedia project in JSON format, or NULL on error.}
\usage{
wmflabs_get_allinfo(
  article,
  project = "en.wikipedia.org",
  links_in_count_redirects = FALSE
)
}
\arguments{
\item{article}{The article to search}

\item{project}{The Wikimedia project, defaults en.wikipedia.org}

\item{links_in_count_redirects}{If infotype==links, if redirects are included or not}
}
\value{
The number of visits
}
\description{
wmflabs_get_allinfo
Obtains information about an article in the Wikimedia project in JSON format, or NULL on error.
}
\note{
Is important that the article be don't a redirection: with "prose" infotype the function gets information of the target article, but with "articleinfo" and "links" the information is about the redirection.
}
\author{
Angel F. Zazo, Departament of Computer Science and Automatics, University of Salamanca
}
